requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
notdiamond>=1.0.0
pydantic>=2.0.0 